........
  r4269 | Lajos | 2006-12-14 11:37:46 -0600 (Thu, 14 Dec 2006) | 1 line
  
  LicenseHandler checks for existence of license.props before building jar file. Removed trap for Exception from GuiLicenseHandler
........


git-svn-id: file:///home/projects/mule/scm@4270 bf997673-6b11-0410-b953-e057580c5b09

+++ branches/mule-1.3/modules/boot/src/main/java/org/mule/modules/boot/LicenseHandler.java
@@ -20,6 +20,7 @@ import java.io.IOException;
 import java.util.jar.JarEntry;
 import java.util.jar.JarFile;
 import java.util.jar.JarOutputStream;
+import org.mule.util.ClassUtils;
 
 /**
  * This class has methods for displaying the EULA and saving the
@@ -180,6 +181,16 @@ public class LicenseHandler
      */
     public void saveLicenseAck(String licenseType, String licenseVersion) throws Exception
     {
+        if (ClassUtils.getResource("META-INF/mule/license.props", MuleBootstrap.class) != null) {
+            // We already have the license.props file from somewhere, so
+            // forget about saving the jar file
+            // 
+            // This case should ONLY be for when the GUI installer saved the
+            // license.props already, and forgot to check for its existence
+            // prior to calling this method.
+            return;
+        }
+
         File muleLib = new File(muleHome, "lib/mule");
         File tempJar = createAckJarFile(licenseType, licenseVersion);
 
@@ -199,6 +210,9 @@ public class LicenseHandler
         // Now we have a directory to create the jar to, so let's rename 
         // the temporary one
         File newJarFile = new File(muleLib, ackJarName);
+        if (newJarFile.exists())
+            throw new Exception("Unable to rename temporary jar to " + newJarFile.getAbsolutePath() + " a file with this name already exists!");
+
         if (!tempJar.renameTo(newJarFile))
             throw new Exception("Unable to rename temporary jar to " + newJarFile.getAbsolutePath());
     }
@@ -252,10 +266,6 @@ public class LicenseHandler
         } finally {
             if (fis != null) {
                 try { fis.close(); } catch (Exception e) { }
-            }
-
-            if (newJar != null) {
-                try { newJar.close(); } catch (Exception e) { }
             }
         }
